working but not freeing

+++ src/thread_ops/thread_ops.h
@@ -6,7 +6,7 @@
 /*   By: mimeyer <mimeyer@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/02 11:21:35 by mimeyer           #+#    #+#             */
-/*   Updated: 2026/05/31 21:30:57 by mimeyer          ###   ########.fr       */
+/*   Updated: 2026/06/08 00:08:19 by mimeyer          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,6 +17,6 @@
 # include "../types/types.h"
 
 int	compile(t_coder *coder, t_dongle **dongles, t_log *log, size_t *data);
-int	wait_log(t_log *log, size_t time_to_pass, int idx, char *arg);
+int	wait_log(t_log *log, uint64_t time_to_pass, int idx, char *arg);
 
 #endif
